vision-rag/pyproject.toml
[project]
name = "vision-rag"
version = "0.1.0"
description = "Visual RAG AI system"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery==5.4.0",
    "fastapi==0.115.6",
    "httpx==0.28.1",
    "minio==7.2.12",
    "numpy==1.26.4",
    "pdf2image==1.17.0",
    "pillow==11.1.0",
    "prometheus-client==0.21.1",
    "prometheus-fastapi-instrumentator==7.0.2",
    "psycopg2-binary==2.9.10",
    "pydantic==2.10.4",
    "pydantic-settings==2.7.1",
    "python-multipart==0.0.20",
    "qdrant-client==1.12.1",
    "redis[hiredis]==5.2.1",
    "structlog==24.4.0",
    "tenacity==9.0.0",
    "uvicorn[standard]==0.34.0",
]

[tool.setuptools.packages.find]
include = ["proxy*"]

[project.optional-dependencies]
test = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "respx>=0.22",
    "fakeredis[lua]>=2.25",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["proxy"]
omit = [
    "proxy/tasks.py",   # Celery worker — requires broker
    "proxy/main.py",    # App entrypoint with lifespan — tested via integration
]

[tool.coverage.report]
show_missing = true
fail_under = 65

[tool.ruff]
target-version = "py312"
line-length = 120
src = ["proxy", "tests"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
    "RUF", # ruff-specific
]
ignore = [
    "E501",   # line length (handled by formatter)
    "B008",   # function call in default argument (FastAPI Depends pattern)
    "RUF012", # mutable class variable (pydantic models)
]

[tool.ruff.lint.isort]
known-first-party = ["proxy"]
